fix(review): address PR #843 review feedback
Address review comments from @dealako:

- campaign.interface.ts + campaign.constants.ts: move
  CampaignPlatformOption and CampaignGoalOption interfaces
  from constants to shared/interfaces (per @dealako)
- campaign-proxy.service.ts: strip ::ffff: prefix before
  SSRF IP pattern matching to block IPv4-mapped IPv6 bypass
  (per @dealako)
- campaign-proxy.service.ts: remove raw Google Ads error
  fallbacks from extractGadsErrorMessage to prevent leaking
  internal details to frontend (per @dealako)
- planning-tab.component.ts: add takeUntilDestroyed to
  lookupHubSpotUtm and createHubSpotUtm subscriptions
  (per @dealako)
- campaign.service.ts: import CAMPAIGN_JOB_POLL_INTERVAL_MS
  from shared constants instead of local redeclaration
  (per @dealako)

LFXV2-2031

Resolves 6 review threads.
